Add files via upload
I couldn't get any updates to work after the initial connection to github. Here's a version that reflects the note below from Slack. I think I just need to look at it with fresh eyes tomorrow.

"I did PCA on the variables and reduced the 13 variables to three components following the directions in a DataCamp lesson called "Reducing Dimensionality with Principal Component Analysis." I'm stuck on the last step -- building a logistic regression with the four components I isolated (it's on page 24 out of 30 of the slides for that lesson). Does that seem like the right direction? I'm worried that the components don't have names that match the original variables 1:1."